Rename namespaces to fit folder structure
