+++ src/Chirp.Core/Entities/Cheep.cs
@@ -1,6 +1,6 @@
 using System.ComponentModel.DataAnnotations;
 
-namespace Core.Entities;
+namespace Chirp.Core.Entities;
 
 public class Cheep
 {
