+++ src/Chirp.Infrastructure/Extensions/CheepExtensions.cs
@@ -1,17 +1,14 @@
-using Core.Entities;
-using Core.Interfaces;
-namespace Infrastructure.Services;
+using Chirp.Core.Entities;
+
+namespace Chirp.Infrastructure.Extensions;
 
 public static class QueryableExtensions
 {
-
     private static readonly int _pageSize = 32;
 
     public static IEnumerable<T> Paginate<T>(this IEnumerable<T> source, int page)
     {
-        return
-        source.Skip(Math.Max(0, page - 1) * _pageSize)
-        .Take(_pageSize);
+        return source.Skip(Math.Max(0, page - 1) * _pageSize).Take(_pageSize);
     }
 
     public static IEnumerable<Cheep> Search(this IEnumerable<Cheep> source, string searchQuery, Func<Cheep, string> fieldSelector)
