+++ src/Chirp.Web/Pages/Public.cshtml
@@ -1,5 +1,5 @@
 ﻿@page "/"
-@model Web.Pages.PublicModel
+@model Chirp.Web.Pages.PublicModel
 @{
     Layout = "Shared/_Layout";
