+++ src/Chirp.Web/Pages/Search.cshtml
@@ -1,5 +1,5 @@
 @page "/search"
-@model Web.Pages.SearchModel
+@model Chirp.Web.Pages.SearchModel
 @{
     Layout = "Shared/_Layout";
