+++ src/Chirp.Web/Pages/UserTimeline.cshtml
@@ -1,5 +1,5 @@
 @page "/{author}"
-@model Web.Pages.UserTimelineModel
+@model Chirp.Web.Pages.UserTimelineModel
 @{
     Layout = "Shared/_Layout";
     var routeName = HttpContext.GetRouteValue("author");
